Update Array to 1.8.7

+++ kernel/common/array.rb
@@ -427,6 +427,28 @@ class Array
     self
   end
 
+  # When invoked with a block, yields all combinations of length n of elements
+  # from ary and then returns ary itself. Even though the implementation makes
+  # no guarantees about the order in which the combinations are yielded, we copy MRI.
+  # When invoked without a block, returns an enumerator object instead.
+  def combination(num)
+    num = Type.coerce_to num, Fixnum, :to_int
+    return to_enum(:combination, num) unless block_given?
+    return self unless (0..size).include? num
+    # Implementation note: slightly tricky.
+                                             # Example: self = 1..7, num = 3
+    picks = (0...num).to_a                   # picks start at 0, 1, 2
+    max = ((size-num)...size).to_a           # max (index for a given pick) is [4, 5, 6]
+    pick_max_pairs = picks.zip(max).reverse  # pick_max_pairs = [[2, 6], [1, 5], [0, 4]]
+    lookup = pick_max_pairs.find(Proc.new{return self})
+    loop do
+      yield values_at(*picks)
+      move = lookup.each{|pick, max| picks[pick] < max}.first
+      new_index = picks[move] + 1
+      picks[move...num] = (new_index...(new_index+num-move)).to_a
+    end
+  end
+
   # Returns a copy of self with all nil elements removed
   def compact
     out = dup
@@ -449,6 +471,20 @@ class Array
     Ruby.primitive :array_concat
 
     concat Type.coerce_to(other, Array, :to_ary)
+  end
+
+  # Calls block for each element repeatedly n times or forever if none
+  # or nil is given. If a non-positive number is given or the array is empty,
+  # does nothing. Returns nil if the loop has finished without getting interrupted.
+  def cycle(n = nil, &block)
+    return to_enum(:cycle, n) unless block_given?
+    if n.nil?
+      loop(&block)
+    else
+      n = Type.coerce_to n, Fixnum, :to_int
+      n.times{each(&block)}
+    end
+    nil
   end
 
   # Removes all elements from self that are #== to the given object.
@@ -497,9 +533,9 @@ class Array
   end
 
   # Deletes every element from self for which block evaluates to true
-  #
-  # NOTE: This method is overridden by lib/1.8.7 or lib/1.9.
   def delete_if(&block)
+    return to_enum :delete_if unless block_given?
+
     key = Undefined
     i = to_iter
     while i.next
@@ -517,9 +553,9 @@ class Array
   # Passes each index of the Array to the given block
   # and returns self.  We re-evaluate @total each time
   # through the loop in case the array has changed.
-  #
-  # NOTE: This method is overridden by lib/1.8.7 or lib/1.9.
   def each_index
+    return to_enum :each_index unless block_given?
+
     i = to_iter
     while i.next
       yield i.index
@@ -670,16 +706,18 @@ class Array
   end
 
   # Recursively flatten any contained Arrays into an one-dimensional result.
-  def flatten
-    dup.flatten! || self
+  # The optional level argument determines the level of recursion to flatten
+  def flatten(level=-1)
+    dup.flatten!(level) || self
   end
 
   # Flattens self in place as #flatten. If no changes are
   # made, returns nil, otherwise self.
-  def flatten!
+  # The optional level argument determines the level of recursion to flatten
+  def flatten!(level=-1)
+    level = Type.coerce_to(level, Integer, :to_int)
     ret, out = nil, []
-
-    ret = recursively_flatten(self, out)
+    ret = recursively_flatten(self, out, level)
     replace(out) if ret
     ret
   end
@@ -710,11 +748,19 @@ class Array
   end
 
   # Returns the index of the first element in the Array
-  # for which elem == obj is true or nil.
-  def index(obj)
+  # for which elem == obj is true or nil. If a block is
+  # given instead of an argument, returns first object
+  # for which block is true. Returns nil if no match is found.
+  def index(obj=Undefined)
     i = to_iter
-    while i.next
-      return i.index if i.item == obj
+    if obj.equal? Undefined
+      while i.next
+        return i.index if yield(i.item)
+      end
+    else
+      while i.next
+        return i.index if i.item == obj
+      end
     end
     nil
   end
@@ -902,18 +948,49 @@ class Array
   end
 
   # Removes and returns the last element from the Array.
-  def pop
-    return nil if empty?
+  def pop(many=Undefined)
+    if many.equal? Undefined
+      return nil if empty?
 
-    @total -= 1
-    index = @start + @total
+      @total -= 1
+      index = @start + @total
 
-    elem = @tuple.at(index)
-    @tuple.put(index,nil)
+      elem = @tuple.at(index)
+      @tuple.put(index,nil)
 
-    reallocate_shrink()
+      reallocate_shrink()
 
-    elem
+      elem
+    else
+      many = Type.coerce_to(many, Fixnum, :to_int)
+      raise ArgumentError, "negative array size" if many < 0
+
+      first = size - many
+      first = 0 if first < 0
+      slice!(first..size).to_a
+    end
+  end
+
+  # Returns an array of all combinations of elements from all arrays.
+  # The length of the returned array is the product of the length of
+  # ary and the argument arrays
+  def product(*arg)
+    # Implementation notes: We build an enumerator for all the combinations
+    # by building it up successively using "inject" and starting from a trivial enumerator.
+    # It would be easy to have "product" yield to a block but the standard
+    # simply returns an array, so you'll find a simple call to "to_a" at the end.
+    #
+    trivial_enum = Enumerator.new_with_block{|yielder| yielder.yield [] }
+    [self, *arg].map{|x| Type.coerce_to(x, Array, :to_ary)}.
+      inject(trivial_enum) do |enum, array|
+        Enumerator.new_with_block do |yielder|
+          enum.each do |partial_product|
+            array.each do |obj|
+              yielder.yield partial_product + [obj]
+            end
+          end
+        end
+    end.to_a
   end
 
   # Appends the given object(s) to the Array and returns
@@ -939,17 +1016,17 @@ class Array
   # Returns a new Array by removing items from self for
   # which block is true. An Array is also returned when
   # invoked on subclasses. See #reject!
-  #
-  # NOTE: This method is overridden by lib/1.8.7 or lib/1.9.
   def reject(&block)
+    return to_enum :reject unless block_given?
+
     Array.new(self).reject!(&block) || self
   end
 
   # Equivalent to #delete_if except that returns nil if
   # no changes were made.
-  #
-  # NOTE: This method is overridden by lib/1.8.7 or lib/1.9.
   def reject!(&block)
+    return to_enum :reject! unless block_given?
+
     was = length
     self.delete_if(&block)
 
@@ -990,9 +1067,9 @@ class Array
 
   # Goes through the Array back to front and yields
   # each element to the supplied block. Returns self.
-  #
-  # NOTE: This method is overridden by lib/1.8.7 or lib/1.9.
   def reverse_each
+    return to_enum :reverse_each unless block_given?
+
     i = to_reverse_iter
     while i.rnext
       yield i.item
@@ -1002,28 +1079,65 @@ class Array
 
   # Returns the index of the last element in the Array
   # for which elem == obj is true.
-  def rindex(obj)
-    i = to_reverse_iter
-    while i.rnext
-      return i.index if i.item == obj
+  # If a block is given instead of an argument,
+  # returns last object for which block is true.
+  def rindex(obj=Undefined)
+    if obj.equal? Undefined
+      i = to_reverse_iter
+      while i.rnext
+        return i.index if yield(i.item)
+      end
+    else
+      i = to_reverse_iter
+      while i.rnext
+        return i.index if i.item == obj
+      end
     end
     nil
+  end
+
+  # Choose a random element, or the random n elements, from the array.
+  # If the array is empty, the first form returns nil, and the second
+  # form returns an empty array.
+  def choice(n=Undefined)
+    return at(rand(size)) if n.equal? Undefined
+
+    n = Type.coerce_to(n, Fixnum, :to_int)
+    raise ArgumentError, "negative array size" if n < 0
+
+    n = size if n > size
+    result = Array.new(self)
+
+    n.times do |i|
+      r = i + rand(size - i)
+      result.tuple.swap(i,r)
+    end
+
+    result[n..size] = []
+    result
   end
 
   # Removes and returns the first element in the
   # Array or nil if empty. All other elements are
   # moved down one index.
-  def shift
-    return nil if @total == 0
+  def shift(n=Undefined)
+    if n.equal? Undefined
+      return nil if @total == 0
 
-    obj = @tuple.at @start
-    @tuple.put @start, nil
-    @start += 1
-    @total -= 1
+      obj = @tuple.at @start
+      @tuple.put @start, nil
+      @start += 1
+      @total -= 1
 
-    reallocate_shrink()
+      reallocate_shrink()
 
-    obj
+      obj
+    else
+      n = Type.coerce_to(n, Fixnum, :to_int)
+      raise ArgumentError, "negative array size" if n < 0
+
+      slice!(0, n).to_a
+    end
   end
 
   # Deletes the element(s) given by an index (optionally with a length)
@@ -1053,6 +1167,20 @@ class Array
     end
 
     out
+  end
+
+  # Returns a new array with elements of this array shuffled.
+  def shuffle
+    dup.shuffle!
+  end
+
+  # Shuffles elements in self in place.
+  def shuffle!
+    size.times do |i|
+      r = i + rand(size - i)
+      @tuple.swap(i,r)
+    end
+    self
   end
 
   # Returns a new Array created by sorting elements in self. Comparisons
